refactor(server,worker): extract app builder pattern for composable configurations
Introduce ServerAppBuilder and WorkerAppBuilder that allow downstream
crates to compose custom server and worker setups with additional
routes, event listeners, migrations, and background tasks.

Extension points on ServerAppBuilder:
- .auth() — custom auth backend factory
- .routes() — additional API routes (billing, etc.)
- .event_listener() — custom event listeners for observability
- .migration() — post-migrate callbacks for custom DB schemas
- .background_task() — custom background tasks with ServerContext

The old server::run() is kept as a thin backward-compatible wrapper.
Both main.rs binaries updated to use the builder directly.
